+++ server-plugins/notification-resources/src/index.ts
@@ -35,24 +35,9 @@ import core, {
 import notification, { EmailNotification, Notification, NotificationStatus } from '@anticrm/notification'
 import { getResource } from '@anticrm/platform'
 import type { TriggerControl } from '@anticrm/server-core'
+import { extractTx } from '@anticrm/server-core'
 import { getUpdateLastViewTx } from '@anticrm/server-notification'
 import view, { HTMLPresenter, TextPresenter } from '@anticrm/view'
-
-const extractTx = (tx: Tx): Tx => {
-  if (tx._class === core.class.TxCollectionCUD) {
-    const ctx = tx as TxCollectionCUD<Doc, AttachedDoc>
-    if (ctx.tx._class === core.class.TxCreateDoc) {
-      const create = ctx.tx as TxCreateDoc<AttachedDoc>
-      create.attributes.attachedTo = ctx.objectId
-      create.attributes.attachedToClass = ctx.objectClass
-      create.attributes.collection = ctx.collection
-      return create
-    }
-    return ctx.tx
-  }
-
-  return tx
-}
 
 /**
  * @public