+++ server-plugins/tags-resources/src/index.ts
@@ -14,7 +14,6 @@
 //
 
 import core, {
-  AttachedDoc,
   Class,
   Doc,
   DocumentQuery,
@@ -23,30 +22,13 @@ import core, {
   Hierarchy,
   Ref,
   Tx,
-  TxCollectionCUD,
   TxCreateDoc,
   TxCUD,
   TxProcessor,
   TxRemoveDoc
 } from '@anticrm/core'
-import { TriggerControl } from '@anticrm/server-core'
+import { extractTx, TriggerControl } from '@anticrm/server-core'
 import tags, { TagElement, TagReference } from '@anticrm/tags'
-
-const extractTx = (tx: Tx): Tx => {
-  if (tx._class === core.class.TxCollectionCUD) {
-    const ctx = tx as TxCollectionCUD<Doc, AttachedDoc>
-    if (ctx.tx._class === core.class.TxCreateDoc) {
-      const create = ctx.tx as TxCreateDoc<AttachedDoc>
-      create.attributes.attachedTo = ctx.objectId
-      create.attributes.attachedToClass = ctx.objectClass
-      create.attributes.collection = ctx.collection
-      return create
-    }
-    return ctx.tx
-  }
-
-  return tx
-}
 
 /**
  * @public